Criando o Método Responsável por Buscar o Carrinho de Compra do Usuário e a View

+++ src/Andreitoledo.GeekShopping.Web/Models/CartHeaderViewModel.cs
@@ -9,6 +9,6 @@
         
         public string CouponCode { get; set; }
 
-        public double PurchaseAmount { get; set; }
+        public decimal PurchaseAmount { get; set; }
     }
 }
